Rebrand as ChickReomte, harden remote sessions, and refresh documentation

+++ code/client/rule/shell/link.go
@@ -3,26 +3,29 @@ package shell
 import (
 	"io"
 	"os"
+	"sync"
 
 	"github.com/lwch/logging"
-	"github.com/lwch/natpass/code/client/conn"
-	"github.com/lwch/natpass/code/network"
-	"github.com/lwch/natpass/code/utils"
 	"golang.org/x/text/encoding/simplifiedchinese"
 	"google.golang.org/protobuf/proto"
+	"org.mutantcat.chickreomte/code/client/conn"
+	"org.mutantcat.chickreomte/code/network"
+	"org.mutantcat.chickreomte/code/utils"
 )
 
 // Link shell link
 type Link struct {
-	parent *Shell
-	id     string // link id
-	target string // target id
-	remote *conn.Conn
+	closeOnce sync.Once
+	parent    *Shell
+	id        string // link id
+	target    string // target id
+	remote    *conn.Conn
 	// in remote
 	pid    int
 	stdin  io.WriteCloser
 	stdout io.ReadCloser
 	// runtime
+	statsMu    sync.RWMutex
 	sendBytes  uint64
 	recvBytes  uint64
 	sendPacket uint64
@@ -36,26 +39,48 @@ func (link *Link) GetID() string {
 
 // GetBytes get send and recv bytes
 func (link *Link) GetBytes() (uint64, uint64) {
+	link.statsMu.RLock()
+	defer link.statsMu.RUnlock()
 	return link.recvBytes, link.sendBytes
 }
 
 // GetPackets get send and recv packets
 func (link *Link) GetPackets() (uint64, uint64) {
+	link.statsMu.RLock()
+	defer link.statsMu.RUnlock()
 	return link.recvPacket, link.sendPacket
+}
+
+func (link *Link) recordSent(n uint64) {
+	link.statsMu.Lock()
+	defer link.statsMu.Unlock()
+	link.sendBytes += n
+	link.sendPacket++
+}
+
+func (link *Link) recordReceived(n uint64) {
+	link.statsMu.Lock()
+	defer link.statsMu.Unlock()
+	link.recvBytes += n
+	link.recvPacket++
 }
 
 // Close close link
 func (link *Link) Close(send bool) {
-	link.onClose()
-	p, err := os.FindProcess(link.pid)
-	if err == nil {
-		p.Kill()
-	}
-	if send {
-		link.remote.SendDisconnect(link.target, link.id)
-	}
-	link.parent.remove(link.id)
-	link.remote.ChanClose(link.id)
+	link.closeOnce.Do(func() {
+		link.onClose()
+		if link.pid > 0 {
+			p, err := os.FindProcess(link.pid)
+			if err == nil {
+				p.Kill()
+			}
+		}
+		if send {
+			link.remote.SendDisconnect(link.target, link.id)
+		}
+		link.parent.remove(link.id)
+		link.remote.ChanClose(link.id)
+	})
 }
 
 // Forward forward data
@@ -74,8 +99,7 @@ func (link *Link) remoteRead() {
 			return
 		}
 		data, _ := proto.Marshal(msg)
-		link.recvBytes += uint64(len(data))
-		link.recvPacket++
+		link.recordReceived(uint64(len(data)))
 		switch msg.GetXType() {
 		case network.Msg_shell_resize:
 			size := msg.GetSresize()
@@ -119,16 +143,14 @@ func (link *Link) localRead() {
 		logging.Debug("link %s on shell %s read from local %d bytes",
 			link.id, link.parent.Name, n)
 		send := link.remote.SendShellData(link.target, link.id, data)
-		link.sendBytes += send
-		link.sendPacket++
+		link.recordSent(send)
 	}
 }
 
 // SendData send data
 func (link *Link) SendData(data []byte) {
 	send := link.remote.SendShellData(link.target, link.id, data)
-	link.sendBytes += send
-	link.sendPacket++
+	link.recordSent(send)
 }
 
 // SendResize send resize message
