+++ code/client/rule/vnc/process/process.go
@@ -9,13 +9,14 @@ import (
 	"net/http"
 	"os"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/gorilla/websocket"
 	"github.com/lwch/logging"
-	"github.com/lwch/natpass/code/client/rule/vnc/vncnetwork"
-	"github.com/lwch/natpass/code/utils"
 	"google.golang.org/protobuf/proto"
+	"org.mutantcat.chickreomte/code/client/rule/vnc/vncnetwork"
+	"org.mutantcat.chickreomte/code/utils"
 )
 
 const (
@@ -25,11 +26,30 @@ const (
 
 // Process process
 type Process struct {
-	pid         int
+	// Keep the atomic PID aligned on 32-bit platforms.
+	pid         int64
+	initOnce    sync.Once
+	closeOnce   sync.Once
+	done        chan struct{}
+	captureMu   sync.Mutex
 	srv         *http.Server
 	chWrite     chan *vncnetwork.VncMsg
 	chImage     chan *vncnetwork.ImageData
 	chClipboard chan *vncnetwork.ClipboardData
+}
+
+func (p *Process) doneChan() <-chan struct{} {
+	p.initOnce.Do(func() { p.done = make(chan struct{}) })
+	return p.done
+}
+
+func (p *Process) send(msg *vncnetwork.VncMsg) bool {
+	select {
+	case p.chWrite <- msg:
+		return true
+	case <-p.doneChan():
+		return false
+	}
 }
 
 func (p *Process) listenAndServe() (uint16, error) {
@@ -65,9 +85,14 @@ func (p *Process) ws(w http.ResponseWriter, r *http.Request) {
 	}
 	defer conn.Close()
 	defer p.Close()
+	go func() {
+		<-p.doneChan()
+		conn.Close()
+	}()
 	var wg sync.WaitGroup
 	wg.Add(2)
 	go func() {
+		defer p.Close()
 		defer utils.Recover("ws read")
 		defer wg.Done()
 		for {
@@ -83,18 +108,32 @@ func (p *Process) ws(w http.ResponseWriter, r *http.Request) {
 			}
 			switch msg.GetXType() {
 			case vncnetwork.VncMsg_capture_data:
-				p.chImage <- msg.GetData()
+				select {
+				case p.chImage <- msg.GetData():
+				case <-p.doneChan():
+					return
+				}
 			case vncnetwork.VncMsg_clipboard_event:
-				p.chClipboard <- msg.GetClipboard()
+				select {
+				case p.chClipboard <- msg.GetClipboard():
+				case <-p.doneChan():
+					return
+				}
 			default:
 			}
 		}
 	}()
 	go func() {
 		defer utils.Recover("ws write")
+		defer p.Close()
 		defer wg.Done()
 		for {
-			msg := <-p.chWrite
+			var msg *vncnetwork.VncMsg
+			select {
+			case msg = <-p.chWrite:
+			case <-p.doneChan():
+				return
+			}
 			data, err := proto.Marshal(msg)
 			if err != nil {
 				continue
@@ -110,7 +149,11 @@ func (p *Process) ws(w http.ResponseWriter, r *http.Request) {
 }
 
 func (p *Process) kill() {
-	ps, _ := os.FindProcess(p.pid)
+	pid := atomic.LoadInt64(&p.pid)
+	if pid <= 0 {
+		return
+	}
+	ps, _ := os.FindProcess(int(pid))
 	if ps != nil {
 		ps.Kill()
 	}
@@ -118,45 +161,50 @@ func (p *Process) kill() {
 
 // Close close process
 func (p *Process) Close() {
-	if p.srv != nil {
-		p.srv.Close()
-	}
-	if p.chImage != nil {
-		close(p.chImage)
-		p.chImage = nil
-	}
-	if p.chClipboard != nil {
-		close(p.chClipboard)
-		p.chClipboard = nil
-	}
-	if p.chWrite != nil {
-		close(p.chWrite)
-		p.chWrite = nil
-	}
-	p.kill()
+	p.doneChan()
+	p.closeOnce.Do(func() {
+		close(p.done)
+		if p.srv != nil {
+			p.srv.Close()
+		}
+		p.kill()
+	})
 }
 
 // Capture capture desktop image
 func (p *Process) Capture(timeout time.Duration) (*image.RGBA, error) {
+	p.captureMu.Lock()
+	defer p.captureMu.Unlock()
+	if timeout <= 0 {
+		timeout = 3 * time.Second
+	}
+	timer := time.NewTimer(timeout)
+	defer timer.Stop()
 	var msg vncnetwork.VncMsg
 	msg.XType = vncnetwork.VncMsg_capture_req
-	p.chWrite <- &msg
-	trans := func(data *vncnetwork.ImageData) *image.RGBA {
-		img := image.NewRGBA(image.Rect(0, 0, int(data.GetWidth()), int(data.GetHeight())))
-		copy(img.Pix, data.GetData())
-		// dumpImage(img)
-		return img
+	select {
+	case p.chWrite <- &msg:
+	case <-timer.C:
+		return nil, errors.New("capture timeout")
+	case <-p.doneChan():
+		return nil, net.ErrClosed
 	}
-	if timeout > 0 {
-		select {
-		case data := <-p.chImage:
-			return trans(data), nil
-		case <-time.After(timeout):
-			return nil, errors.New("timeout")
+	select {
+	case data := <-p.chImage:
+		if !data.GetOk() {
+			return nil, fmt.Errorf("capture failed: %s", data.GetMsg())
 		}
-	} else {
-		data := <-p.chImage
-		return trans(data), nil
+		w, h := uint64(data.GetWidth()), uint64(data.GetHeight())
+		if w == 0 || h == 0 || w > 32768 || h > 32768 || w*h*4 != uint64(len(data.GetData())) {
+			return nil, errors.New("invalid capture dimensions")
+		}
+		img := image.NewRGBA(image.Rect(0, 0, int(w), int(h)))
+		copy(img.Pix, data.GetData())
+		return img, nil
+	case <-timer.C:
+		return nil, errors.New("capture timeout")
+	case <-p.doneChan():
+		return nil, net.ErrClosed
 	}
 }
